+++ macros/src/impl_conf.rs
@@ -7,6 +7,9 @@ use super::CONF_OPTION;
 pub fn impl_conf(ast: &syn::DeriveInput) -> TokenStream {
     let name_conf = format_ident!("{}{CONF_OPTION}", ast.ident);
     let name = &ast.ident;
+    let help_heading = name.to_string().replace("Config", "");
+
+            // #[clap(long, group = "export", help_heading = "Export")]
     let fields = match &ast.data {
         syn::Data::Struct(data) => match &data.fields {
             syn::Fields::Named(named) => &named.named,
@@ -25,6 +28,7 @@ pub fn impl_conf(ast: &syn::DeriveInput) -> TokenStream {
         let mandatory = quote! {
             #[arg(long)]
             #[serde()]
+            #[clap(help_heading = #help_heading)]
             #(#attrs)*
             pub #field_name: Option<#ty>,
         };
